+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+- Added a settings page in the control panel.
 - Webhook URLs can now be set to environment variables or Twig code. ([#18](https://github.com/craftcms/webhooks/issues/18))
 - Webhooks can now send PUT requests. ([#21](https://github.com/craftcms/webhooks/issues/21))